Turned PushNotificationResponse into an interface and added a private default implementation. Documented the interface.
